r/virtual_machine and friends: WIP large refactor
To make device workflow a bit more sane, and to un-crowd the namespace,
I've started to move helpers to their own packages, in addition to the
device sub-resources. This has had a bit of a waterfall effect though,
so there is probably a bit more refactoring that's going to be necessary
to get all of the helpers into their own packages in the respective
trees.

The project here does not build right now currently, I'm just saving my
work for the evening.