More refactors
The provider should build now as all concerned identifiers have been
renamed.

There will probably be more refactoring as things move on here,
depending on what else we need in the device sub-resources, and then
later on, to keep things consistent (it doesn't make sense that we've
moved some of the helpers down into internal packages and just left the
others, and there is possibly some lingering duplication that could be
fixed).