r/distributed_virtual_switch: Fix folder identifier
Looks like it was legit go to re-assign folder and use it as an
identifier in scope if there was no functions actually tied to it after
the fact. Fixed this for now.

In relation to this, we might refactor he folder package to a different
name as it seems to be colliding more often than not with variable
names, as "folder" is shorter and more prone to being used as a variable
than the rest of the package names have been. May be a TODO, we will see
how annoying it gets.