r/virtual_machine: Some state tweaks on refresh and schema gen fixes
* Refresh now purnes unknown devices from state after trying to locate
every device it can find. This helps with diffs - if a device drifts
and the config version can't be found anymore, it gets removed
completely, which will show up in a plan now as the orphaned device
getting fully removed and the config version being re-added from
scratch.
* Removed some logging we aren't going to keep.
* Also moved sub-schema generation for virtual device to functions.
Using variables was having the unintended side effect of merging right
into the package-level maps, which was ultimately leading to key
conflicts and possibly other trouble. Using functions cleanly creates
new maps every time to work off of.