r/virtual_machine: Use local device connection logic for virtual NICs
AssignController in VirtualDeviceList doesn't assign PCI devices
properly right now, so we have to bake our own connection logic for now.
I've put in an issue with in govmomi to see what can be done about it,
if anything, but I've just de-coupled us for now.

Virtual NICs seem to start at unit number 7 on the VM's PCI controller.
The default device set has the VMCI controller at unit number 17
attached to the PCI bus, so we allow a maximum of 10 virtual devices to
be assigned to the PCI controller before returning an error.