r/virtual_machine: WIP refactor of virtual device workflows
The virtual device workflows are changing: the arbitrary index is going
away, with network_interface and cdrom devices now working off lists and
having their unit numbers, and hence order, inferred (in the case of
network interfaces), or just being locked down to a single device (in
the case of CDROM drives). Disks are getting a new unit_number attribute
which directly translates to where on the SCSI bus it gets placed.
Further to this, all SCSI controllers are being locked down to a certain
type, with the entire bus filled by pre-creating all four slots.
Currently we allow VMware PV and LSI logic, with no plans to change that
(the other controllers right now are legacy).

This split in approaches to handling each device class has caused enough
of a divergence that using Subresource as it currently stands is proving
to be more trouble than it's worth. More than likely its role will
diminish as I continue de-coupling things.

I'm committing now before I start seriously gutting the object. The code
here builds but it crashes on tests.