r/virtual_machine: Don't crash on unknown disk backing types
There is a chance that someone may be adding virtual disks that are not
file-based, such as in the case of RDM disks or what not. We don't
support these right now but we can't assume that every disk we see will
be of a correct type, so fail gracefully if it isn't.