r/virtual_machine: Test and fixes for SCSI bus change
This update now allows you to change the devices on the SCSI bus to
different device types - an intended feature but something that was
untested so far.

Also, simplified most of the tests to use a shorter form of the disk
sub-resource - for the first disk (device unit_number 0), the
configuration directive can be omitted. We check the unit numbers on the
diff validation.