r/virtual_machine: Disk handling and state migration/import updates
* We now only check the controllers dictated by scsi_controller_count
for disks.
* A scsi_controller_count setting now needs to be supplied to import
when requiring more than 1 controller. The input to import is now JSON.
* Fixed a bug with handling device unit numbers for disks past the first
controller - units were not being converted correctly to their >14
numbers, which was causing spurious diffs.
* There is now additional logic for properly interpreting the
pre-migration state of wait_for_guest_net into a proper
wait_for_guest_net_timeout value (-1 for false, 5 for true).
* State migration now tries to determine the SCSI controller count based
on the number of disks, and by checking the bus number/unit number of
each disk that is currently in state. This should help with cases where
a VM has more than 15 disks added.
* Tests have been added for adding disks at unit numbers higher than 14,
which is how the unit translation bug was uncovered.
* Updated the docs with all of the new information. This pretty much
removes the entire part of reading a disk diff, as it's no longer
invalid and is not any more complex than the rest of the resource's
diff.