r/virtual_machine: Test for IPv6-only and extra docs
This extra coverage covers an IPv6-only scenario, which is kind of
possible in that you can elect to have IPv4 setup via DHCP, and
configure static IPv6. In this case where the IPv4 network is not
routeable and you have not set up DHCP, you more than likely need to
configure the guest network waiter for a long enough period that it gets
past the initial boot's DHCP timeout (about 5 minutes on Ubuntu).

Note that this is not an issue per se with the waiter - during this
timeout period, VMware tools has not been started yet, and the IPv6
address possibly has not been configured, so the whole VM is stuck
service wise until the timeout runs its course, so the only 2 courses of
action here are to either configure the waiter for a period where the
data will be available, or to turn it off.